+++ frontend/app/src/people/main/BountyModal/index.ts
