+++ frontend/app/src/people/main/EditUserModal/EditUserDesktopView.tsx
@@ -6,7 +6,7 @@ import FocusedView from '../focusView';
 import { formConfig } from './config';
 import { observer } from 'mobx-react-lite';
 
-export const EditUserDesctopView = observer(() => {
+export const EditUserDesktopView = observer(() => {
   const {canEdit, closeHandler, person, modals} = useUserEdit();
   return (
     <Modal
