+++ frontend/app/src/people/main/EditUserModal/index.tsx
@@ -1,21 +1,17 @@
-import React, { useEffect } from 'react'
 import { useIsMobile } from 'hooks'
-import { EditUserDesctopView } from './EditUserDesctopView'
-import { EditUserMobileView } from './EditUserMobileView'
-import { useStores } from 'store'
 import { observer } from 'mobx-react-lite'
+import React from 'react'
+import { useStores } from 'store'
+import { EditUserDesktopView } from './EditUserDesktopView'
+import { EditUserMobileView } from './EditUserMobileView'
 
 export const EditUserModal = observer(() => {
   const isMobile = useIsMobile();
   const {modals} = useStores();
-  useEffect(() => {
-    console.log('EditUserModal')
-    console.log(modals.userEditModal)
 
-  }, [modals.userEditModal])
   if (!modals.userEditModal) {
     return null
   }
 
-  return <>{isMobile ? <EditUserMobileView /> : <EditUserDesctopView />}</>
+  return <>{isMobile ? <EditUserMobileView /> : <EditUserDesktopView />}</>
 })
